TST: Moved I/O tests to IO
